+++ packages/experience/src/Layout/FocusedAuthPageLayout/index.module.scss
