+++ packages/experience/src/Layout/FocusedAuthPageLayout/index.tsx
@@ -22,7 +22,13 @@ type Props = {
   readonly authOptionsLink: TextLinkProps;
 };
 
-const IdentifierPageLayout = ({
+/**
+ * FocusedAuthPageLayout Component
+ *
+ * This layout component is designed for focused authentication pages that serve as the first screen
+ * for specific auth methods, such as identifier sign-in, identifier-register, and single sign-on landing pages.
+ */
+const FocusedAuthPageLayout = ({
   children,
   pageMeta,
   title,
@@ -52,4 +58,4 @@ const IdentifierPageLayout = ({
   );
 };
 
-export default IdentifierPageLayout;
+export default FocusedAuthPageLayout;
