+++ packages/experience/src/pages/SingleSignOnEmail/SingleSignOnForm/index.module.scss
@@ -16,4 +16,8 @@
     margin-left: _.unit(0.5);
     margin-top: _.unit(-3);
   }
+
+  .terms {
+    margin-bottom: _.unit(4);
+  }
 }
